Restructured code

+++ types.h
@@ -7,8 +7,8 @@ typedef uint64_t chunk;
 #define ONE 1ULL
 typedef uint32_t meter;
 typedef uint16_t place;
-typedef uint16_t agent;
+typedef uint32_t agent;
 typedef uint32_t penny;
-typedef uint16_t edge;
+typedef uint32_t edge;
 typedef uint32_t id;
 typedef float dist;
